+++ services/subscriber/app/Providers/BindServiceProvider.php
@@ -5,9 +5,7 @@ namespace App\Providers;
 use App\Adapters\CurrencyRateAdapter;
 use App\Interfaces\Adapters\CurrencyRateAdapterInterface;
 use App\Interfaces\Repositories\SubscriberRepositoryInterface;
-use App\Interfaces\Services\RabbitMQServiceInterface;
 use App\Repositories\SubscriberRepository;
-use App\Services\RabbitMQService;
 use Illuminate\Support\ServiceProvider;
 
 class BindServiceProvider extends ServiceProvider
@@ -43,6 +41,5 @@ class BindServiceProvider extends ServiceProvider
     private function registerServices(): void
     {
         $this->app->bind(CurrencyRateAdapterInterface::class, CurrencyRateAdapter::class);
-        $this->app->bind(RabbitMQServiceInterface::class, RabbitMQService::class);
     }
 }