+++ CHANGELOG.md
@@ -5,10 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] - 2024-07-10
+
+### Added
+- move send emails command to mailer service
+- added unsubscribe route
+
+## [1.3.2] - 2024-07-03
+
+### Added
+- create mailer microservices
+
 ## [1.3.1] - 2024-07-02
 
 ### Added
-- create 2 microservices
+- create 2 microservices (currency, subscriber)
 
 ## [1.2.1] - 2024-06-27